Refactor to use dblite over better-sqlite3
The better-sqlite3 package continued to cause node-gyp errors on systems that do not contain a working python installation or with permission errors. I have replaced the better-sqlite3 package with dblite as it contains an implementation that should relieve these errors.
I tried sql.js before this and although functional, it is out-of-date and uses depreciated methods making it less than ideal.